move modes/params to modes/paramdef
