+++ specparam/modes/__init__.py
@@ -1,5 +1,5 @@
 """Functionality related to definining fit modes."""
 
 from .mode import Mode
-from .params import ParamDefinition
 from .check import check_modes
+from .paramdef import ParamDefinition
